Homework wk2_d02
