--- new_grocery_list.csv
@@ -1,1 +1,0 @@
-oranges,kale,bread
